feat: register market simulations
