+++ x/market/simulation/errors.go
@@ -1,0 +1,5 @@
+package simulation
+
+const (
+	errTextInvalidMarketKey = "invalid market key %X"
+)
