+++ x/market/simulation/market.go
@@ -29,3 +29,22 @@ func SimulateMsgAddMarket(
 		return simtypes.NoOpMsg(types.ModuleName, msg.Type(), "AddMarket simulation not implemented"), nil, nil
 	}
 }
+
+// SimulateMsgResolveMarket simulates the resolve market flow
+func SimulateMsgResolveMarket(
+	ak types.AccountKeeper,
+	bk types.BankKeeper,
+	k keeper.Keeper,
+) simtypes.Operation {
+	return func(r *rand.Rand, app *baseapp.BaseApp, ctx sdk.Context, accs []simtypes.Account, chainID string,
+	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
+		simAccount, _ := simtypes.RandomAcc(r, accs)
+		msg := &types.MsgResolveMarket{
+			Creator: simAccount.Address.String(),
+		}
+
+		// TODO: Handling the ResolveMarket simulation
+
+		return simtypes.NoOpMsg(types.ModuleName, msg.Type(), "ResolveMarket simulation not implemented"), nil, nil
+	}
+}
